requirements.txt
chromadb
sentence-transformers
Pillow
numpy
python-dotenv
tqdm
google-generativeai
pdf2image
